Minor fixes, and working on reinstantiating the client
