+++ frontend/pages/account/register.jsx
@@ -1,75 +1,34 @@
-import { useRouter } from 'next/router';
-import { useForm } from 'react-hook-form';
-import { yupResolver } from '@hookform/resolvers/yup';
-import * as Yup from 'yup';
+import React from "react";
 
-// custom components
-import { Link } from '../../components/Login/index';
-import { Layout } from '../../components/Login/Layout/index';
-import { register_account } from '../../components/Stomp';
+// client imports
+import { Provider } from "react-redux";
+import { CookiesProvider } from "react-cookie";
+import { SWRConfig } from 'swr'
 
-export default Register;
+// store
+import store from "../../stores/store";
 
-function Register() {
-    const router = useRouter();
+// custom imports
+import Registration from "../../components/Login/RegisterUser/Registration";
 
-    // form validation rules 
-    const validationSchema = Yup.object().shape({
-        firstName: Yup.string()
-            .required('First Name is required'),
-        lastName: Yup.string()
-            .required('Last Name is required'),
-        username: Yup.string()
-            .required('Username is required'),
-        password: Yup.string()
-            .required('Password is required')
-            .min(6, 'Password must be at least 6 characters')
-    });
-    const formOptions = { resolver: yupResolver(validationSchema) };
-
-    // get functions to build form with useForm() hook
-    const { register, handleSubmit, formState } = useForm(formOptions);
-    const { errors } = formState;
-
-    function onSubmit(user) {
-        register_account(user);
-        router.push('login');
-    }
+const fetcher = (...args) => fetch(...args).then((res) => res.json());
+console.log("Fetcher", fetcher)
+export default function Register() {
 
     return (
-        <Layout>
-            <div className="card">
-                <h4 className="card-header">Register</h4>
-                <div className="card-body">
-                    <form onSubmit={handleSubmit(onSubmit)}>
-                        <div className="form-group">
-                            <label>First Name</label>
-                            <input name="firstName" type="text" {...register('firstName')} className={`form-control ${errors.firstName ? 'is-invalid' : ''}`} />
-                            <div className="invalid-feedback">{errors.firstName?.message}</div>
-                        </div>
-                        <div className="form-group">
-                            <label>Last Name</label>
-                            <input name="lastName" type="text" {...register('lastName')} className={`form-control ${errors.lastName ? 'is-invalid' : ''}`} />
-                            <div className="invalid-feedback">{errors.lastName?.message}</div>
-                        </div>
-                        <div className="form-group">
-                            <label>Username</label>
-                            <input name="username" type="text" {...register('username')} className={`form-control ${errors.username ? 'is-invalid' : ''}`} />
-                            <div className="invalid-feedback">{errors.username?.message}</div>
-                        </div>
-                        <div className="form-group">
-                            <label>Password</label>
-                            <input name="password" type="password" {...register('password')} className={`form-control ${errors.password ? 'is-invalid' : ''}`} />
-                            <div className="invalid-feedback">{errors.password?.message}</div>
-                        </div>
-                        <button disabled={formState.isSubmitting} className="btn btn-primary">
-                            {formState.isSubmitting && <span className="spinner-border spinner-border-sm mr-1"></span>}
-                            Register
-                        </button>
-                        <Link href="/account/login" className="btn btn-link">Cancel</Link>
-                    </form>
-                </div>
-            </div>
-        </Layout>
+        <div>
+            <SWRConfig value={{
+                fetcher: fetcher,
+                errorRetryCount: 10,
+                refreshInterval: 250
+            }}>
+                <CookiesProvider>
+                    <Provider store={store}>
+                        <Registration/>
+                    </Provider>
+                </CookiesProvider>
+            </SWRConfig>
+        </div>
     );
 }
+
