+++ frontend/pages/index.js
@@ -14,7 +14,8 @@ import clientPromise from '../util/mongodb';
 
 // API fetcher for SWR global config
 //const fetcher = (...args: Parameters<typeof fetch>) => fetch(...args).then((res) => res.json())
-const fetcher = (...args) => fetch(...args).then((res) => res.json())
+const fetcher = (...args) => fetch(...args).then((res) => res.json());
+console.log("Index Fetcher", fetcher)
 
 export default function Home({ isConnected }) {
   return (
